Property collation and edge-vertex link support
The commit has following changes.

1. collation of a property
--------------------------
The values with differing collations can not be collated. This means
that we will not able to compare, sort, or order values of properties
with the same name if they have different collations in different
elements or different labels.  This restricts the property's usage in
graph table query. Hence the collation of all the properties with the
same name needs to be the same. Enforce this.

Often the collation of a property reference is required before
graph_table is rewritten. For example, assign_query_collation() is
called during transformSelectStmt() which is called before calling
rewriteGraphTable(). Hence the collation of a property needs to be
accessible before any property reference is accessed. Hence the
collation of a property is stored in pg_propgraph_property.

Note: This is not explicitly specified in Section 9.15, "Consistency
check of a tabular property graph descriptor", syntax rule 4.c.iii.2 of
SQL/PGQ standard.

2. collation of source and destination keys of an edge
------------------------------------------------------
If collations of edge key and the corresponding referenced key differ,
an edge may end up being adjacent to undesired vertex. Prohibit such a
case when key columns are explicitly specified.

When edge and vertex keys are derived from the foreign key constraint,
we do not check collations again since the constraint itself ensures
one-to-many edge-to-vertex mapping.

Note: This is not specified in Section 9.14, "Creation of an edge table
descriptor".

3. Edge-vertex link quals
-----------------------------------------
When creating an edge element make sure that there exists an equality
operators that can be used to match vertex and edge keys. Fail the DDL if
such an operator does not exist. Use the same equality operators to build
vertex-edge quals in build_edge_vertex_link_quals(). Also add a
dependency of the edge on the equality operator so that it can not be
dropped without dropping the edge.

Note for reviewers:
The code in ATAddForeignKeyConstraint() looks up three different
equality operators and it's difficult to separate the code looking up
just PK=FK operators. So could not move some common code to a function
which can be used at both places.

4. Tests
--------
Adds tests for above items.

The collation tests, test that

a. the collation is correctly set for the columns projected by
GRAPH_TABLE clause.

b. quals corresponding to the various WHERE clause in GRAPH_TABLE have
correct collation set

c. quals corresponding to the edge-vertex links have correct collation
set.

Note: These tests combined with tests in collate.sql and other collation
specific tests indicate that we have covered all scenarios testing
collations in the context of GRAPH_TABLE. More collation specific tests
may be added as required.

Author: Ashutosh Bapat
Reviewed by: Junwang Zhou