Remove the need to check if plan is valid from ExecutorStart hooks
For testing using delay_execution, a new hook
ExecutotStartCachedPlan_hook is added.  This hook allows the
delay_execution module to block the execution of the cached plan to
allow a concurrent session to modify the objects referenced in the
plan, which is then detected when the locks are taken on prunable
relations in ExecutorStart().